Time: 11 ms (72.28%), Space: 6.4 MB (17.29%) - LeetHub

+++ 0326-power-of-three/0326-power-of-three.cpp
@@ -8,6 +8,6 @@ public:
 
         if(n == 0) return false;
         
-        return double_is_int((log10(n) / (double)log10(3)));
+        return double_is_int((log10(n) / log10(3)));
     }
 };
